relaxng.c: cast to allow compilation with sun studio 11
Sun Studio 11 was failing to compile relaxng.c due to a type mismatch
in a ternary operator used to provide an argument to xmlRngPErr.  It
seems that sos11 cc is more pedantic about this part of the C99
standard.  GCC issues a warning but doesn't fail by default.

After casting the string literal "nothing" to (const xmlChar *) sos11
cc is happy.  GCC no longer issues a warning either.

Another warning was cleaned up with an identical cast.