fix runtests to use pthreads support for various Unix platforms
The runtests program currently fails with

    Specific platform thread support not detected

on HP-UX, AIX and other Unix systems which do not match the conditional

    #if defined(linux) || defined(__sun) || defined(__APPLE_CC__)

It is silly to try to enumerate all systems which use pthreads in a conditional
like this. I am attaching a patch (against git master) that rewrites the cpp
conditional structure so that pthreads is used if HAVE_PTHREAD_H is defined,
and moves that section of code down below the Win32 and BeOS cases so that
native thread libraries are used preferentially in those two cases.