build: Use pkg-config to find liblzma in preference to AC_CHECK_LIB
This means that liblzma’s Libs.private will be included in LZMA_LIBS if
linking the libraries statically, ensuring that there are no undefined
symbol errors from liblzma’s private libraries.

If pkg-config isn’t installed, or if liblzma.pc couldn’t be found, fall
back to using AC_CHECK_LIB as before. This will cause static linking to
fail, but that’s not a regression.

This does not introduce a compile time dependency on pkg-config.

https://bugzilla.gnome.org/show_bug.cgi?id=711026