win32\VC10\config.h and VS 2015
--047d7bacb4ee71848a052d7c673a
Content-Type: text/plain; charset=UTF-8

The config.h file in win32\VC10 is incompatible with VS 2015 because VS
2015 provides an implementation of snprintf and prohibits using the
preprocessor to define it. Therefor an ifdef check is needed around that
definition in VC10\config.h. Here is a patch that is compatible with the
change that we made to Chromium's copy: