trio*h: Decorate symbols for export
Include xmlexports.h and define TRIO_PUBLIC to be XMLPUBFUN by default.
This will ensure that we export the symbols as needed when we build in
TRIO support.

Decroate the prototypes in the headers with TRIO_PUBLIC (or
TRIO_STRING_PUBLIC, which is in turn defined to TRIO_PUBLIC) so that we
are able to use this to export those symbols as well.

On Visual Studio, one is not allowed to only decorate the function
declarations with __declspec(dllexport) or __declspec(dllimport) without
decorating the prototypes with the same things (but, doing the vice-versa
is allowed).